Move `@/util/vue` to `@/composables`
