+++ app/gui2/src/components/ComponentBrowser.vue
@@ -7,6 +7,9 @@ import { default as DocumentationPanel } from '@/components/DocumentationPanel.v
 import GraphVisualization from '@/components/GraphEditor/GraphVisualization.vue'
 import SvgIcon from '@/components/SvgIcon.vue'
 import ToggleIcon from '@/components/ToggleIcon.vue'
+import { useApproach } from '@/composables/animation'
+import { useEvent, useResizeObserver } from '@/composables/events'
+import type { useNavigator } from '@/composables/navigator'
 import { useGraphStore } from '@/stores/graph'
 import type { RequiredImport } from '@/stores/graph/imports'
 import { useProjectStore } from '@/stores/project'
@@ -17,9 +20,6 @@ import { tryGetIndex } from '@/util/data/array'
 import type { Opt } from '@/util/data/opt'
 import { allRanges } from '@/util/data/range'
 import { Vec2 } from '@/util/data/vec2'
-import { useApproach } from '@/util/vue/animation'
-import { useEvent, useResizeObserver } from '@/util/vue/events'
-import type { useNavigator } from '@/util/vue/navigator'
 import type { SuggestionId } from 'shared/languageServerTypes/suggestions'
 import type { ContentRange, ExprId } from 'shared/yjsModel'
 import { computed, nextTick, onMounted, ref, watch, type ComputedRef, type Ref } from 'vue'
