+++ app/gui2/src/components/ExecutionModeSelector.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import SvgIcon from '@/components/SvgIcon.vue'
-import { useEvent } from '@/util/vue/events'
+import { useEvent } from '@/composables/events'
 import { ref } from 'vue'
 
 const props = defineProps<{ modes: string[]; modelValue: string }>()
