+++ app/gui2/src/components/GraphEditor/dragging.ts
@@ -1,3 +1,4 @@
+import { useApproach } from '@/composables/animation'
 import { injectGraphSelection } from '@/providers/graphSelection'
 import { useGraphStore } from '@/stores/graph'
 import { partitionPoint } from '@/util/data/array'
@@ -5,7 +6,6 @@ import type { Opt } from '@/util/data/opt'
 import { Rect } from '@/util/data/rect'
 import { Vec2 } from '@/util/data/vec2'
 import theme from '@/util/theme.json'
-import { useApproach } from '@/util/vue/animation'
 import { iteratorFilter } from 'lib0/iterator'
 import type { ExprId } from 'shared/yjsModel'
 import { computed, markRaw, ref, watchEffect, type ComputedRef, type WatchStopHandle } from 'vue'
