+++ app/gui2/src/components/visualizations/HistogramVisualization.vue
@@ -1,9 +1,9 @@
 <script lang="ts">
 import SvgIcon from '@/components/SvgIcon.vue'
+import { useEvent } from '@/composables/events'
 import { getTextWidth } from '@/util/measurement'
 import { defineKeybinds } from '@/util/shortcuts'
 import { VisualizationContainer, useVisualizationConfig } from '@/util/visualizationBuiltins'
-import { useEvent } from '@/util/vue/events'
 import { computed, ref, watch, watchEffect, watchPostEffect } from 'vue'
 
 export const name = 'Histogram'
