+++ app/gui2/src/components/visualizations/ScatterplotVisualization.vue
@@ -1,9 +1,9 @@
 <script lang="ts">
 import SvgIcon from '@/components/SvgIcon.vue'
+import { useEvent } from '@/composables/events'
 import { useVisualizationConfig } from '@/providers/visualizationConfig'
 import { getTextWidth } from '@/util/measurement'
 import { VisualizationContainer, defineKeybinds } from '@/util/visualizationBuiltins'
-import { useEvent } from '@/util/vue/events'
 import { computed, ref, watch, watchEffect, watchPostEffect } from 'vue'
 
 export const name = 'Scatter Plot'
