+++ app/gui2/src/components/widgets/ListWidget.vue
@@ -1,10 +1,10 @@
 <script lang="ts">
 import SvgIcon from '@/components/SvgIcon.vue'
+import { useRaf } from '@/composables/animation'
+import { useEvent } from '@/composables/events'
 import { useAppClass } from '@/providers/appClass'
 import { Range } from '@/util/data/range'
 import { Vec2 } from '@/util/data/vec2'
-import { useRaf } from '@/util/vue/animation'
-import { useEvent } from '@/util/vue/events'
 import { uuidv4 } from 'lib0/random'
 import { nextTick } from 'process'
 import { computed, ref, shallowReactive, watchEffect, watchPostEffect } from 'vue'
