+++ app/gui2/src/composables/animation.ts
@@ -1,6 +1,6 @@
 /** @file Vue composables for running a callback on every frame, and smooth interpolation. */
 
-import { watchSourceToRef } from '@/util/vue/reactivity'
+import { watchSourceToRef } from '@/util/reactivity'
 import { onScopeDispose, proxyRefs, ref, watch, type WatchSource } from 'vue'
 
 const rafCallbacks: { fn: (t: number, dt: number) => void; priority: number }[] = []
