+++ app/gui2/src/composables/doubleClick.ts
