+++ app/gui2/src/composables/navigator.ts
@@ -1,9 +1,9 @@
 /** @file A Vue composable for panning and zooming a DOM element. */
 
+import { useApproach } from '@/composables/animation'
+import { PointerButtonMask, useEvent, usePointer, useResizeObserver } from '@/composables/events'
 import { Rect } from '@/util/data/rect'
 import { Vec2 } from '@/util/data/vec2'
-import { useApproach } from '@/util/vue/animation'
-import { PointerButtonMask, useEvent, usePointer, useResizeObserver } from '@/util/vue/events'
 import { computed, proxyRefs, ref, type Ref } from 'vue'
 
 function elemRect(target: Element | undefined): Rect {
