+++ app/gui2/src/providers/graphNavigator.ts
@@ -1,5 +1,5 @@
+import { useNavigator } from '@/composables/navigator'
 import { createContextStore } from '@/providers'
-import { useNavigator } from '@/util/vue/navigator'
 
 export type GraphNavigator = ReturnType<typeof injectFn>
 export { injectFn as injectGraphNavigator, provideFn as provideGraphNavigator }
