+++ app/gui2/src/providers/graphSelection.ts
@@ -1,7 +1,7 @@
+import type { NavigatorComposable } from '@/composables/navigator'
+import { useSelection } from '@/composables/selection'
 import { createContextStore } from '@/providers'
 import type { Rect } from '@/util/data/rect'
-import type { NavigatorComposable } from '@/util/vue/navigator'
-import { useSelection } from '@/util/vue/selection'
 import type { ExprId } from 'shared/yjsModel'
 
 const SELECTION_BRUSH_MARGIN_PX = 6
