+++ app/gui2/src/util/shortcuts.ts
@@ -1,4 +1,4 @@
-import { isMacLike } from '@/util/vue/events'
+import { isMacLike } from '@/composables/events'
 
 /** All possible modifier keys. */
 export type ModifierKey = keyof typeof RAW_MODIFIER_FLAG
