+++ app/gui2/stories/GraphNode.story.vue
@@ -103,4 +103,3 @@ const SetupStory = createSetupComponent((app) => {
     </template>
   </Story>
 </template>
-@/util/data/vec2 @/util/data/rect
