+++ app/gui2/stories/UserDefinedVisualization.story.vue
@@ -104,4 +104,3 @@ const mockFsWrapperProps = computed(() => ({
     </template>
   </Story>
 </template>
-@/util/data/vec2
